+++ modules/libmod_net/libmod_net.c
@@ -44,51 +44,51 @@ int64_t libmod_net_open( INSTANCE *my, int64_t *params ) {
     char * addr = ( char * ) string_get( params[ 2 ] );
     void * r = ( void * ) net_open( params[ 0 ], params[ 1 ], *addr ? addr : NULL, params[ 3 ] );
     string_discard( params[ 2 ] );
-    return ( int64_t ) r;
+    return ( int64_t )( intptr_t ) r;
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_wait( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_wait( ( List * ) params[ 0 ], params[ 1 ], ( List * ) params[ 2 ] );
+    return ( int64_t ) net_wait( ( List * )( intptr_t ) params[ 0 ], params[ 1 ], ( List * )( intptr_t ) params[ 2 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_send( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_send( ( _net * ) params[ 0 ], ( const void * ) params[ 1 ], ( size_t ) params[ 2 ] );
+    return ( int64_t ) net_send( ( _net * )( intptr_t ) params[ 0 ], ( const void * )( intptr_t ) params[ 1 ], ( size_t ) params[ 2 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_recv( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_recv( ( _net * ) params[ 0 ], ( void * ) params[ 1 ], ( size_t ) params[ 2 ] );
+    return ( int64_t ) net_recv( ( _net * )( intptr_t ) params[ 0 ], ( void * )( intptr_t ) params[ 1 ], ( size_t ) params[ 2 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_close( INSTANCE *my, int64_t *params ) {
-    net_close( ( _net * ) params[ 0 ] );
+    net_close( ( _net * )( intptr_t ) params[ 0 ] );
     return 0;
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_getavailablebytes( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_getavailablebytes( ( _net * ) params[ 0 ] );
+    return ( int64_t ) net_getavailablebytes( ( _net * )( intptr_t ) params[ 0 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_getReceiveBufferSize( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_getReceiveBufferSize( ( _net * ) params[ 0 ] );
+    return ( int64_t ) net_getReceiveBufferSize( ( _net * )( intptr_t ) params[ 0 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_getremoteaddr( INSTANCE *my, int64_t *params ) {
     int64_t r = 0;
-    char * ip = net_getremoteaddr( ( _net * ) params[ 0 ] );
+    char * ip = net_getremoteaddr( ( _net * )( intptr_t ) params[ 0 ] );
     if ( ip ) {
         r = string_new( ip ) ;
         free( ip );
@@ -102,13 +102,13 @@ int64_t libmod_net_getremoteaddr( INSTANCE *my, int64_t *params ) {
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_is_new_connection( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_is_new_connection( ( _net * ) params[ 0 ] );
+    return ( int64_t ) net_is_new_connection( ( _net * )( intptr_t ) params[ 0 ] );
 }
 
 /* --------------------------------------------------------------------------- */
 
 int64_t libmod_net_is_message_incoming( INSTANCE *my, int64_t *params ) {
-    return ( int64_t ) net_is_message_incoming( ( _net * ) params[ 0 ] );
+    return ( int64_t ) net_is_message_incoming( ( _net * )( intptr_t ) params[ 0 ] );
 }
 
 /* --------------------------------------------------------------------------- */
